+++ tests/test_clean.py
@@ -2,7 +2,7 @@ from pytest_mock import MockerFixture
 
 from .base_test import BaseTest
 from .mockers import (assert_success, launch_command, mock_input_returning_y,
-                      rewrite_branch_layout_file)
+                      read_branch_layout_file, rewrite_branch_layout_file)
 from .mockers_github import mock_github_token_for_domain_none
 
 
@@ -50,9 +50,10 @@ class TestClean(BaseTest):
 
         launch_command('clean')
 
+        assert read_branch_layout_file() == "master\n    bar\n    foo\n    moo\n"
+
         expected_status_output = (
             """
-            Warning: sliding invalid branches: bar2, foo2, moo2, mars out of the branch layout file
               master *
               |
               o-bar (untracked)